Add files via upload
Revised MWSegEval_ToolBox.mlapp to fix bug (i.e., revised code so that the ‘Segmentation’ switch remains in either the ‘automatic’ or ‘not automatic’ after requesting the ‘input forward and reconstructed models for processing’ task. Previously, if the switch was in the ‘automatic’ position, it would change state to the ‘not automatic’ state after the input model task was requested). -D.Kurrant